% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_timestamp}
\alias{add_timestamp}
\title{Add timestamp onto end of api url to prevent caching issues}
\usage{
add_timestamp(api_path)
}
\arguments{
\item{api_path}{string; the API url to retrieve data from}
}
\description{
Add timestamp onto end of api url to prevent caching issues
}
