% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_types.R
\name{area_types}
\alias{area_types}
\title{Area types}
\usage{
area_types()
}
\value{
A data frame of area type ids and their descriptions
}
\description{
Outputs a data frame of area type ids and their descriptions
}
