% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areaTypes.R
\name{areaTypes}
\alias{areaTypes}
\title{Area types}
\usage{
areaTypes()
}
\value{
A data frame of area type ids and their descriptions
}
\description{
Outputs a data frame of area type ids and their descriptions
}

