% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{areatypes_by_indicators}
\alias{areatypes_by_indicators}
\title{IndicatorID by AreaTypeID}
\format{A data frame with 3593 rows and 2 variables:
\describe{
  \item{IndicatorID}{ID for indicator in Fingertips database, see  indicators()}
  \item{AreaTypeID}{ID for area types used in Fingertips database, see area_types()}
  ...
}}
\source{
\url{https://fingertips.phe.org.uk/}
}
\usage{
areatypes_by_indicators
}
\description{
A dataset containing indicator ids and the area types that data exists for
}
\keyword{datasets}
