% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_types.R
\name{category_types}
\alias{category_types}
\title{Category types}
\usage{
category_types()
}
\value{
A data frame of category type ids and their descriptions
}
\description{
Outputs a data frame of category type ids, their name (along with a short name)
}
\seealso{
\code{\link{indicators}} for indicator lookups,
  \code{\link{profiles}} for profile lookups and
  \code{\link{deprivation_decile}} for deprivation decile lookups and
  \code{\link{area_types}} for area type lookups and
  \code{\link{indicator_areatypes}} for indicators by area types lookups and
  \code{\link{indicators_unique}} for unique indicatorids and their names

Other lookup functions: \code{\link{area_types}},
  \code{\link{deprivation_decile}},
  \code{\link{indicator_areatypes}},
  \code{\link{indicator_metadata}},
  \code{\link{indicators_unique}},
  \code{\link{indicators}}, \code{\link{profiles}}
}
