% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprivationDecile.R
\name{deprivationDecile}
\alias{deprivationDecile}
\title{Get deprivation deciles}
\usage{
deprivationDecile(AreaTypeID = 102, Year = 2015)
}
\value{
A lookup table providing deprivation decile and area code
}
\description{
Outputs a data frame allocating deprivation decile to  area code
}

