% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fingertips_deframe}
\alias{fingertips_deframe}
\title{fingertips_deframe}
\usage{
fingertips_deframe(data)
}
\arguments{
\item{data}{list whose first item is a vector of names, and second item is a
list. Items 1 and 2 must be equal length}
}
\description{
mimic tibble::deframe() without needing to import the function
}
