% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fingertips_endpoint}
\alias{fingertips_endpoint}
\title{Get the default fingertips API endpoint}
\usage{
fingertips_endpoint()
}
\value{
A character string with the HTTP URL of the Fingertips API
}
\description{
Get the default fingertips API endpoint
}
