% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fingertips_ensure_api_available}
\alias{fingertips_ensure_api_available}
\title{Check if the given Fingertips API endpoint is available}
\usage{
fingertips_ensure_api_available(endpoint = fingertips_endpoint())
}
\arguments{
\item{endpoint}{string, the API base URL to check}
}
\value{
\code{TRUE} if the API is available, otherwise \code{stop()} is called.
}
\description{
Check if the given Fingertips API endpoint is available
}
