% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fingertips_proxy_settings}
\alias{fingertips_proxy_settings}
\title{fingertips_proxy_settings}
\usage{
fingertips_proxy_settings()
}
\value{
A character string with the proxy settings
}
\description{
determines which proxy settings are used
}
