% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhancements.R
\name{fingertips_redred}
\alias{fingertips_redred}
\title{Red significance and red trend}
\usage{
fingertips_redred(ProfileID = NULL, AreaTypeID = 102,
  Comparator = "England")
}
\arguments{
\item{ProfileID}{Numeric vector, id of profiles of interest. Indicator polarity can vary between profiles therefore if using
one of the comparison fields it is recommended to complete this field as well as IndicatorID. If IndicatorID is populated,
ProfileID can be ignored or must be the same length as IndicatorID (but can contain NAs).}

\item{AreaTypeID}{Numeric vector, the Fingertips ID for the area type;
default is 102}

\item{Comparator}{String, either "England" or "Parent" to determine which field to compare the spot value significance to}
}
\description{
Filters data returned by the fingertips_data function for values for areas that are trending significantly
worse and the spot value is significantly worse than the England value in the latest year of that indicator
#' @return A data frame of data extracted from the Fingertips API
}
\examples{
\dontrun{
# Returns data for the two selected domains at county and unitary authority geography
reddata <- fingertips_redred(ProfileID = 26, AreaTypeID = 102)}
#' @family data extract functions
}
