% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhancements.R
\name{fingertips_stats}
\alias{fingertips_stats}
\title{High level statistics on Fingertips data}
\usage{
fingertips_stats(proxy_settings = fingertips_proxy_settings())
}
\arguments{
\item{proxy_settings}{string; whether to use Internet Explorer proxy settings
"default" or "none"; default determined automatically.}
}
\value{
A string that summarises the high level statistics of indicators and
  profiles in Fingertips
}
\description{
A sentence that summarises the number of indicators, unique indicators and
profiles
}
\examples{
\dontrun{
# Returns a sentence describing number of indicators and profiles in Fingertips
fingertips_stats()}
}
