% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fingertipsData.R
\name{fingertipsData}
\alias{fingertipsData}
\title{Fingertips data}
\usage{
fingertipsData(IndicatorID = NULL, AreaCode = NULL, DomainID = NULL,
  ProfileID = NULL, AreaTypeID = 102, ParentCode = NULL)
}
\value{
A data frame of data extracted from the Fingertips API
}
\description{
Outputs a data frame of data from [Fingertips]<http://fingertips.phe.org.uk/>
}

