% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fingertipsR.R
\docType{package}
\name{fingertipsR}
\alias{fingertipsR-package}
\alias{fingertipsR}
\title{fingertipsR: A package for extracting the data behind the Fingertips website
(\url{https://fingertips.phe.org.uk/})}
\description{
The fingertipsR package provides two categories of important functions: lookup
and data extract.
}
\section{Lookup functions}{
 The lookup functions are to provide users the
  ability to understand the ID inputs for the data extract functions.
}

\section{Data extract functions}{
 Using ID codes as inputs, the data extract
  functions allow the user to extract data from the Fingertips API.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://fingertips.phe.org.uk}
  \item \url{https://github.com/ropensci/fingertipsR}
  \item \url{https://docs.ropensci.org/fingertipsR/}
  \item Report bugs at \url{https://github.com/ropensci/fingertipsR/issues}
}

}
\author{
\strong{Maintainer}: Annabel Westermann \email{annabel.westermann@dhsc.gov.uk}

Authors:
\itemize{
  \item Sebastian Fox \email{sebsfox@gmail.com} [contributor]
  \item Hadley Nanayakkara \email{hadley.nanayakkara@dhsc.gov.uk} [contributor]
  \item Julian Flowers [contributor]
}

Other contributors:
\itemize{
  \item Simon Thelwall \email{simon.thelwall@ukhsa.gov.uk} (\href{https://orcid.org/0000-0002-0434-2724}{ORCID}) [contributor]
  \item Duncan Garmonsway \email{nacnudus@gmail.com} [contributor]
  \item Carl Ganz \email{carlganz@gmail.com} [contributor]
  \item David Whiting \email{david.whiting@medway.gov.uk} [contributor]
  \item Luke Bradley \email{dev.1b7@outlook.com} [contributor]
  \item Samuel Cutler \email{samuel.cutler@ukhsa.gov.uk} [contributor]
  \item Crown Copyright 2020 [copyright holder]
}

}
