% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_fingertips_api}
\alias{get_fingertips_api}
\title{Retrieve data from a given Fingertips API url}
\usage{
get_fingertips_api(api_path)
}
\arguments{
\item{api_path}{string; the API url to retrieve data from}
}
\description{
Retrieve data from a given Fingertips API url
}
