% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_types.R
\name{indicator_areatypes}
\alias{indicator_areatypes}
\title{Area types by indicator}
\usage{
indicator_areatypes()
}
\value{
A data frame of indicator ids and area type ids
}
\description{
Outputs a data frame of indicator ids and the area type ids that exist for that indicator
}
\seealso{
\code{\link{indicators}} for indicator lookups,
  \code{\link{profiles}} for profile lookups and
  \code{\link{deprivation_decile}} for deprivation decile lookups and
  \code{\link{area_types}} for area type lookups and
  \code{\link{category_types}} for category type lookups

Other lookup functions: \code{\link{area_types}},
  \code{\link{category_types}},
  \code{\link{deprivation_decile}},
  \code{\link{indicator_metadata}},
  \code{\link{indicators}}, \code{\link{profiles}}
}
