% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_metadata.R
\name{indicator_update_information}
\alias{indicator_update_information}
\title{Indicator update information}
\usage{
indicator_update_information(
  IndicatorID,
  ProfileID = NULL,
  proxy_settings = "default",
  path
)
}
\arguments{
\item{IndicatorID}{Integer, id of the indicators of interest}

\item{ProfileID}{Integer (optional), whether to restrict the indicators to a particular profile}

\item{proxy_settings}{string; whether to use Internet Explorer proxy settings
("default") or "none"}

\item{path}{String; Fingertips API address. Function will default to the
correct address}
}
\value{
The date of latest data update for selected indicators
}
\description{
Outputs a data frame which provides a date of when an indicator was last update
}
\examples{
\dontrun{
# Returns metadata for indicator ID 90362 and 1107
indicatorIDs <- c(90362, 1107)
indicator_update_information(indicatorIDs)}
}
