% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liveProfiles.R
\name{liveProfiles}
\alias{liveProfiles}
\title{Live profiles}
\usage{
liveProfiles(profileId = NULL, profileName = NULL)
}
\value{
A data frame of live profile ids and names along with their domain names and ids.
}
\description{
Outputs a data frame of live profiles that data are available for in [Fingertips]<http://fingertips.phe.org.uk/>
}

