% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_types.R
\name{nearest_neighbour_areatypeids}
\alias{nearest_neighbour_areatypeids}
\title{Nearest neighbours area type ids}
\usage{
nearest_neighbour_areatypeids(proxy_settings = fingertips_proxy)
}
\arguments{
\item{proxy_settings}{string; whether to use Internet Explorer proxy settings
("default") or "none"}
}
\value{
table of AreaTypeIDs
}
\description{
Outputs a table of AreaTypeIDs available for the nearest_neighbour function
}
\examples{
\dontrun{
nearest_neighbour_areatypeids()}
}
\seealso{
\code{\link{nearest_neighbours}} to access the geogaphy codes of the
  nearest neighbours for a locality
}
