% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_types.R
\name{nearest_neighbours}
\alias{nearest_neighbours}
\title{Nearest neighbours}
\usage{
nearest_neighbours(
  AreaCode,
  AreaTypeID,
  proxy_settings = fingertips_proxy_settings(),
  path
)
}
\arguments{
\item{AreaCode}{Character vector, ONS area code of area of interest}

\item{AreaTypeID}{AreaTypeID of the nearest neighbours (see
\code{\link{nearest_neighbour_areatypeids}}) for available IDs}

\item{proxy_settings}{string; whether to use Internet Explorer proxy settings
"default" or "none". Setting this manually will decrease runtime;
 default determined automatically.}

\item{path}{String; Fingertips API address. Function will default to the
correct address}
}
\value{
A character vector of area codes
}
\description{
Outputs a character vector of similar areas for a given area.\cr\cr\strong{Details}\cr
\itemize{
\item For new upper tier local authorities (post 4/23) the
\href{https://github.com/NHSDigital/ASC_LA_Peer_Groups}{NHSE Nearest Neighbours
Model} is used.
\item All other local authorities types (lower tier and older upper tier) use the
\href{https://www.cipfastats.net/resources/nearestneighbours/}{CIPFA's
Nearest Neighbours Model}. Older local authority geography types will use
older versions of the model.
\item Similar areas for Clinical Commissioning Groups are based on
\href{https://www.england.nhs.uk/publication/similar-10-ccg-explorer-tool/}{NHS
England's similar CCG explorer tool}.
\item Currently the function does not have the ability to use the method from the
\href{https://www.gov.uk/government/publications/local-authority-interactive-tool-lait}{Children's
services statistical neighbour benchmarking tool}.
}
}
\examples{
\dontrun{
nearest_neighbours(AreaCode = "E38000002", AreaTypeID = 154)}
}
\seealso{
\code{\link{nearest_neighbour_areatypeids}} for the AreaTypeIDs
available for this function

Other lookup functions: 
\code{\link{area_types}()},
\code{\link{category_types}()},
\code{\link{deprivation_decile}()},
\code{\link{indicator_areatypes}()},
\code{\link{indicator_metadata}()},
\code{\link{indicator_order}()},
\code{\link{indicators}()},
\code{\link{indicators_unique}()},
\code{\link{profiles}()}
}
\concept{lookup functions}
