% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_types.R
\name{category_types}
\alias{category_types}
\title{Category types}
\usage{
category_types()
}
\value{
A data frame of area type ids and their descriptions
}
\description{
Outputs a data frame of area type ids, their descriptions, and how they map
to parent area types
}
\seealso{
\code{\link{indicators}} for indicator lookups,
  \code{\link{profiles}} for profile lookups and
  \code{\link{deprivation_decile}} for deprivation decile lookups and
  \code{\link{area_types}} for area type lookups

Other lookup functions: \code{\link{area_types}},
  \code{\link{deprivation_decile}},
  \code{\link{indicator_metadata}},
  \code{\link{indicators}}, \code{\link{profiles}}
}
