% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fingertips_data.R
\name{fingertips_data}
\alias{fingertips_data}
\title{Fingertips data}
\usage{
fingertips_data(IndicatorID = NULL, AreaCode = NULL, DomainID = NULL,
  ProfileID = NULL, AreaTypeID = 102, ParentCode = NULL)
}
\value{
A data frame of data extracted from the Fingertips API
}
\description{
Outputs a data frame of data from [Fingertips]<http://fingertips.phe.org.uk/>
}
