% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhancements.R
\name{fingertips_redred}
\alias{fingertips_redred}
\title{Red significance and red trend}
\usage{
fingertips_redred(Comparator = "England", ...)
}
\arguments{
\item{Comparator}{String, either "England" or "Parent" to determine which field to compare the spot value significance to}

\item{...}{Parameters provided to fingertips_data()}
}
\description{
Filters data returned by the fingertips_data function for values for areas that are trending significantly
worse and the spot value is significantly worse than the England value in the latest year of that indicator
#' @return A data frame of data extracted from the Fingertips API
}
\examples{
\dontrun{
# Returns data for the two selected domains at county and unitary authority geography
reddata <- fingertips_redred(ProfileID = 26, AreaTypeID = 102)}
}
\seealso{
Other data extract functions: \code{\link{fingertips_data}}
}
