% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhancements.R
\name{fingertips_stats}
\alias{fingertips_stats}
\title{High level statistics on Fingertips data}
\usage{
fingertips_stats()
}
\value{
A string that summarises the high level statistics of indicators in
  Fingertips
}
\description{
A sentence that summarises the number of unique indicators, domains and
profiles
}
\examples{
\dontrun{
# Returns a sentence describing number of indicators, domains and profiles in Fingertips
fingertips_stats()}
}
