% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fingertips.R
\docType{package}
\name{fingertipsR}
\alias{fingertipsR}
\alias{fingertipsR-package}
\title{fingertips: A package for extracting the data behind the Fingertips website
(\url{fingertips.phe.gov.uk})}
\description{
The fingertips package provides two categories of important functions: lookup
and data extract.
}
\section{Lookup functions}{
 The lookup functions are to provide users the
  ablity to understand the ID inputs for the data extract functions.
}

\section{Data extract functions}{
 Using ID codes as inputs, the data extract
  functions allow the user to extract data from the Fingertips API.
}

