% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_fingertips_api}
\alias{get_fingertips_api}
\title{Retrieve data from a given Fingertips API url}
\usage{
get_fingertips_api(
  api_path,
  content_type = "text",
  col_types,
  proxy_settings = "default"
)
}
\arguments{
\item{api_path}{string; the API url to retrieve data from}

\item{content_type}{string; "text" or "parsed"}

\item{col_types}{character; vector of column classes}

\item{proxy_settings}{string; whether to use Internet Explorer proxy settings
("default") or "none"}
}
\description{
Retrieve data from a given Fingertips API url
}
\examples{
df <- get_fingertips_api(
  api_path = paste0(
    fingertips_endpoint(),
    "/area/parent_areas?child_area_code=E12000005&parent_area_type_ids=15"))
}
