% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{indicators}
\alias{indicators}
\title{Live profiles}
\usage{
indicators(profileId = NULL, profileName = NULL, DomainID = NULL,
  DomainName = NULL)
}
\value{
A data frame of indicators within a profile or domain.
}
\description{
Outputs a data frame of indicators within a profile or domain
}

