% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{indicators}
\alias{indicators}
\title{Live indicators and the profiles and domains they belong to}
\usage{
indicators(ProfileID = NULL, DomainID = NULL)
}
\arguments{
\item{ProfileID}{Numeric vector, id of profiles of interest}

\item{DomainID}{Numeric vector, id of domains of interest}
}
\value{
A data frame of indicators within a profile or domain.
}
\description{
Outputs a data frame of indicators within a profile or domain
}
\examples{
\dontrun{
# Returns a complete data frame of indicators and their domains and profiles
indicators()

# Returns a data frame of all of the indicators in the Public Health Outcomes Framework
indicators(ProfileID = 19)}
}
\seealso{
\code{\link{area_types}} for area type  and their parent mappings,
  \code{\link{indicator_metadata}} for indicator metadata and
  \code{\link{profiles}} for profile lookups and
  \code{\link{deprivation_decile}} for deprivation decile lookups and
  \code{\link{category_types}} for category lookups and
  \code{\link{indicator_areatypes}} for indicators by area types lookups and
  \code{\link{indicators_unique}} for unique indicatorids and their names

Other lookup functions: \code{\link{area_types}},
  \code{\link{category_types}},
  \code{\link{deprivation_decile}},
  \code{\link{indicator_areatypes}},
  \code{\link{indicator_metadata}},
  \code{\link{indicators_unique}}, \code{\link{profiles}}
}
