% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_types.R
\name{nearest_neighbours}
\alias{nearest_neighbours}
\title{Nearest neighbours}
\usage{
nearest_neighbours(
  AreaCode,
  AreaTypeID,
  measure,
  proxy_settings = fingertips_proxy_settings(),
  path
)
}
\arguments{
\item{AreaCode}{Character vector, ONS area code of area of interest}

\item{AreaTypeID}{AreaTypeID of the nearest neighbours (see
\code{\link{nearest_neighbour_areatypeids}}) for available IDs}

\item{measure}{deprecated. Previously a string; when AreaTypeID = 102 measure
must be either "CIPFA" for CIPFA local authority nearest neighbours or
"CSSN" for Children's services statistical neighbours}

\item{proxy_settings}{string; whether to use Internet Explorer proxy settings
"default" or "none". Setting this manually will decrease runtime;
 default determined automatically.}

\item{path}{String; Fingertips API address. Function will default to the
correct address}
}
\value{
A character vector of area codes
}
\description{
Outputs a character vector of similar areas for given area. Currently returns
similar areas for Clinical Commissioning Groups (old and new) based on
\href{https://www.england.nhs.uk/publication/similar-10-ccg-explorer-tool/}{NHS
England's similar CCG explorer tool} or lower and upper tier local
authorities based on
\href{https://www.cipfastats.net/resources/nearestneighbours/}{CIPFA's
Nearest Neighbours Model} or upper tier local authorities based on
\href{https://www.gov.uk/government/publications/local-authority-interactive-tool-lait}{Children's
services statistical neighbour benchmarking tool}
}
\examples{
\dontrun{
nearest_neighbours(AreaCode = "E38000002", AreaTypeID = 154)}
}
\seealso{
\code{\link{nearest_neighbour_areatypeids}} for the AreaTypeIDs
  available for this function

Other lookup functions: 
\code{\link{area_types}()},
\code{\link{category_types}()},
\code{\link{deprivation_decile}()},
\code{\link{indicator_areatypes}()},
\code{\link{indicator_metadata}()},
\code{\link{indicator_order}()},
\code{\link{indicators_unique}()},
\code{\link{indicators}()},
\code{\link{profiles}()}
}
\concept{lookup functions}
