% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiles.R
\name{profiles}
\alias{profiles}
\title{Live profiles}
\usage{
profiles(profileId = NULL, profileName = NULL)
}
\value{
A data frame of live profile ids and names along with their domain names and ids.
}
\description{
Outputs a data frame of live profiles that data are available for in [Fingertips]<http://fingertips.phe.org.uk/>
}
