% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_indicators.R
\name{select_indicators}
\alias{select_indicators}
\title{Select indicator}
\usage{
select_indicators()
}
\value{
A numeric vector of indicator IDs
}
\description{
Point and click method of selecting indicators and assigning them to object
}
\examples{
\dontrun{
# Opens a browser window allowing the user to select indicators by their name, domain and profile
inds <- select_indicator()}
}
